+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - [feat: support custom quickstart circuitbreak delay time.](https://github.com/Tencent/spring-cloud-tencent/pull/1688)
 - [feat: add delay interface in tsf-example.](https://github.com/Tencent/spring-cloud-tencent/pull/1689)
 - [fix: fix lb configuration on bootstrap step.](https://github.com/Tencent/spring-cloud-tencent/issues/1690)
+- [feat:support fault injection.](https://github.com/Tencent/spring-cloud-tencent/pull/1691)